fix(lingo): fix wrong method resolution for op byte write

+++ Shockky.SourceGeneration.Tests/InstructionGeneratorTests.cs
@@ -62,12 +62,12 @@ public sealed class InstructionGeneratorTests
                 }
                 else if (Immediate <= ushort.MaxValue)
                 {
-                    output.Write(op + 0x40);
+                    output.Write((byte)(op + 0x40));
                     output.Write((ushort)Immediate);
                 }
                 else
                 {
-                    output.Write(op + 0x80);
+                    output.Write((byte)(op + 0x80));
                     output.Write(Immediate);
                 }
             }
